FEAT: Change type if statement. set unsigned int in place of unsigned long.
Mysql return statement number into 4 octet. The long is the same length
that a system pointer. in 64 bit systems, a long is 8 octets in place
of 4

+++ exemple/client_stmt.c
@@ -87,7 +87,7 @@ int main(int argc, char *argv[]) {
 	const char *login;
 	const char *pass;
 	const char *db;
-	unsigned long stmt_id;
+	unsigned int stmt_id;
 	MYSAC_BIND *vars;
 	int n;
 	char *error;
