+++ mysac.h
@@ -263,7 +263,7 @@ typedef struct mysac {
 	enum my_query_st qst;
 	int read_id;
 	MYSAC_RES *res;
-	unsigned long *stmt_id;
+	unsigned int *stmt_id;
 
 	/* the buffer */
 	unsigned int bufsize;
@@ -536,7 +536,7 @@ int mysac_send_query(MYSAC *mysac);
  *
  * @return 0: ok, -1 nok
  */
-int mysac_set_stmt_prepare(MYSAC *mysac, unsigned long *stmt_id, const char *fmt, ...);
+int mysac_set_stmt_prepare(MYSAC *mysac, unsigned int *stmt_id, const char *fmt, ...);
 
 /**
  * Prepare statement
@@ -548,7 +548,7 @@ int mysac_set_stmt_prepare(MYSAC *mysac, unsigned long *stmt_id, const char *fmt
  *
  * @return 0: ok, -1 nok
  */
-int mysac_v_set_stmt_prepare(MYSAC *mysac, unsigned long *stmt_id, const char *fmt, va_list ap);
+int mysac_v_set_stmt_prepare(MYSAC *mysac, unsigned int *stmt_id, const char *fmt, va_list ap);
 
 /**
  * Prepare statement
@@ -559,7 +559,7 @@ int mysac_v_set_stmt_prepare(MYSAC *mysac, unsigned long *stmt_id, const char *f
  *
  * @return 0: ok, -1 nok
  */
-int mysac_s_set_stmt_prepare(MYSAC *mysac, unsigned long *stmt_id, const char *request);
+int mysac_s_set_stmt_prepare(MYSAC *mysac, unsigned int *stmt_id, const char *request);
 
 /**
  * Prepare statement
@@ -571,7 +571,7 @@ int mysac_s_set_stmt_prepare(MYSAC *mysac, unsigned long *stmt_id, const char *r
  *
  * @return 0: ok, -1 nok
  */
-int mysac_b_set_stmt_prepare(MYSAC *mysac, unsigned long *stmt_id, const char *request, int len);
+int mysac_b_set_stmt_prepare(MYSAC *mysac, unsigned int *stmt_id, const char *request, int len);
 
 /**
  * Send sql query command
@@ -598,7 +598,7 @@ int mysac_send_stmt_prepare(MYSAC *mysac);
  *
  * @return 0: ok, -1 nok
  */
-int mysac_set_stmt_execute(MYSAC *mysac, MYSAC_RES *res, unsigned long stmt_id,
+int mysac_set_stmt_execute(MYSAC *mysac, MYSAC_RES *res, unsigned int stmt_id,
                            MYSAC_BIND *values, int nb);
 
 /**
